+++ clang/docs/ReleaseNotes.rst
@@ -156,8 +156,9 @@ CUDA Support in Clang
 X86 Support in Clang
 --------------------
 
-- The x86 intrinsics ``_mm_popcnt_u32`` and ``_mm_popcnt_u64`` may now be used
-  within constexpr expressions.
+- The x86 intrinsics ``_mm_popcnt_u32``, ``_mm_popcnt_u64``, ``_popcnt32``,
+  ``_popcnt64``, ``__popcntd`` and ``__popcntq``  may now be used within
+  constexpr expressions.
 
 Internal API Changes
 --------------------